# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DataClean – Excel/CSV Data Cleaning API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /upload:
    post:
      operationId: uploadFile
      tags: [upload]
      summary: Upload an Excel or CSV file and create a session
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/FileUploadInput"
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
        "400":
          description: Invalid file
...
components:
...
  schemas:
...
    UploadResult:
      type: object
      required: [session_id, sheets, original_filename]
      properties:
        session_id:
          type: string
        sheets:
          type: array
          items:
            type: string
        original_filename:
          type: string
    FileUploadInput:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
...